% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_paths.R
\name{atlantis_examples}
\alias{atlantis_examples}
\title{Helper function to quickly access example files.}
\usage{
atlantis_examples(...)
}
\arguments{
\item{...}{further arguments passed to \verb{[system.file()]}}
}
\description{
Helper function to quickly access example files.
}
\examples{
atlantis_examples("inputs", "tiny.bgm")
}
