% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atlantis_supported_versions}
\alias{atlantis_supported_versions}
\title{Atlantis versions supported by atlantio}
\format{
A data frame, each row includes details about one supported Atlantis
version, currently it includes two columns:
\itemize{
\item version: Atlantis version, formatted as
\code{"<major_version>-<svn_revision>"} (e.g. \code{"3-6722"}),
\item release_date: release date of the version.
}
}
\source{
Generated from \code{inst/raw-data/supported_version.R}.
}
\usage{
atlantis_supported_versions
}
\description{
Atlantis versions supported by atlantio
}
\seealso{
\code{\link[=atlantis_parameter_list]{atlantis_parameter_list()}}
}
\keyword{datasets}
