% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantis_methods.R
\name{create_diet_table}
\alias{create_diet_table}
\title{Create Diet Table}
\usage{
create_diet_table(x, ...)
}
\arguments{
\item{x}{Object containing diet/bio data}

\item{...}{Additional arguments}
}
\value{
A data.frame representing the diet table with predator-prey relationships
}
\description{
Generic function for creating diet tables from model data.
}
\details{
In Atlantis model, the availability of prey is determined by the product of
the prey biomass and the prey choice coefficient \code{pPREY}. The latter is a
proportion (between 0 and 1) of prey biomass actually available to the
predator. The function finds this information in the model and creates the
diet table that is the collection of the pPREY values.

Atlantis distinguishes prey choice values for juveniles and adults. The
distinction exists if species have ontogenetic stages and if the flag
\code{age_structured_prey_XXX} is set to one. Note that for those species the
first mature age group is given by \code{XXX_age_mat}. Therefore there are up to
4 \code{pPREY} values per prey-predator combination:
\itemize{
\item pPREY1XXX1: juvenile-juvenile
\item pPREY1XXX2: juvenile-adult
\item pPREY2XXX1: adult-juvenile
\item pPREY2XXX2: adult-adult
where XXX denotes the code for the prey species. Note that the realised diet
in Atlantis will depend on other factor, for instance the coeficient \code{mum},
assuming it is used by the functional response (see \code{predcase_XXXX}).
}
}
\examples{
my_model <- new_atlantis() |>
  atlantis_load_files(
    c(
      system.file("examples", "inputs", "tiny.bgm", package = "atlantio"),
      system.file("examples", "inputs", "tiny_biol.prm", package = "atlantio"),
      system.file("examples", "inputs", "tiny_groups.csv", package = "atlantio")
    )
  )
}
\references{
\itemize{
\item Asta Audzijonyte, Rebecca Gorton, Isaac Kaplan, Jacob M. Kasper, Elizabeth
A. Fulton, 2025. Atlantis User’s Guide Part I: General Overview, Physics &
Ecology \url{https://research.csiro.au/atlantis/wp-content/uploads/sites/52/2025/05/AtlantisUserGuide_PartI.pdf}
}
}
