% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantis_methods.R
\name{create_foodweb}
\alias{create_foodweb}
\title{Create Food Web for Atlantis objects}
\usage{
create_foodweb(x, ...)
}
\arguments{
\item{x}{An object to create food web from}

\item{...}{Additional arguments passed to methods}
}
\value{
A food web data structure
}
\description{
S7 generic for creating food web data structures from Atlantis objects
}
