% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantis_methods-ts.R
\name{create_time_series}
\alias{create_time_series}
\title{Create Time Series for Atlantis objects}
\usage{
create_time_series(x, ...)
}
\arguments{
\item{x}{An object to create time series from}

\item{...}{Additional arguments passed to methods}
}
\value{
A time series data structure
}
\description{
S7 generic for creating time series data from Atlantis objects
}
\examples{
\dontrun{
my_model <- new_atlantis() |>
  atlantis_load_files(
    c(
      system.file("examples", "inputs", "tiny.bgm", package = "atlantio"),
      system.file("examples", "inputs", "tiny_biol.prm", package = "atlantio"),
      system.file("examples", "inputs", "tiny_groups.csv", package = "atlantio"),
      system.file("examples", "outputs", "output.nc", package = "atlantio")
    )
  )
my_model |> create_time_series(variable = "volume", box = 40)

library(ggplot2)
my_model |>
  create_time_series(variable = "Walleye5_ResN") |>
  ggplot(aes(x = Time, y = value, color = box)) +
  geom_line()
}

}
