% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantis_methods-ts.R
\name{create_time_series_biomass}
\alias{create_time_series_biomass}
\title{Create Time Series Biomass for Atlantis objects}
\usage{
create_time_series_biomass(x, ...)
}
\arguments{
\item{x}{An object to create biomass time series from}

\item{...}{Additional arguments passed to methods}
}
\value{
A biomass time series data structure
}
\description{
S7 generic for creating biomass time series data from Atlantis objects
}
\examples{
\dontrun{
new_atlantis() |>
  atlantis_load_files(
    c(
      system.file("examples", "inputs", "tiny.bgm", package = "atlantio"),
      system.file("examples", "inputs", "tiny_biol.prm", package = "atlantio"),
      system.file("examples", "inputs", "tiny_groups.csv", package = "atlantio"),
      system.file("examples", "outputs", "output.nc", package = "atlantio")
    )
  ) |>
  create_time_series_biomass(group = "YPH")
}

}
