% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{input_files}
\alias{input_files}
\title{Input files}
\format{
A data frame, each row includes details about input files, currently it
includes four columns:
\itemize{
\item name: file name,
\item description: short description,
\item type: either 'input' (model input) or 'output' (model output),
\item file_extension: file extension.
}
}
\usage{
input_files
}
\description{
Input files
}
\keyword{datasets}
