% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_list.R
\name{list_atlantis_parameters}
\alias{list_atlantis_parameters}
\title{List Atlantis parameters for a given Atlantis version}
\usage{
list_atlantis_parameters(version = "3-6722")
}
\arguments{
\item{version}{Atlantis version, formatted as
\code{"<major_version>-<svn_revision>"} (e.g. \code{"3-6722"}). See
\link{atlantis_supported_versions} for the list of supported versions.}
}
\value{
A list of 2 elements:
\itemize{
\item \code{meta}: metadata about the Atlantis code base the parameters were
extracted from (\code{atlantis_source}, \code{atlantis_version}, \code{bm_struct}) and
\code{files}, the list of parameter files declaring the parameters;
\item \code{parameters}: a list where every element describes one parameter with
the following fields: \code{name}, \code{description}, \code{source_file}, \code{value_type},
\code{dimension}, \code{units}, \code{bm_member}, \code{reader} and \code{conditional_on}.
}
}
\description{
Returns the list of Atlantis parameters bundled with the
package for a given Atlantis version. Supported versions are listed in
\link{atlantis_supported_versions}.
}
\examples{
prm_list <- list_atlantis_parameters()
prm_list$meta
prm_list$parameters[[1]]

}
