% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantis_methods-calibrate.R
\name{generate_calibration_table}
\alias{generate_calibration_table}
\title{Generate a calibration table}
\usage{
generate_calibration_table(x, ...)
}
\arguments{
\item{x}{An \code{Atlantis} object.}

\item{...}{Additional arguments passed to methods}
}
\value{
A list of data frames, one per parameter entry, with columns
\code{name}, \code{min}, \code{max}, \code{position}, \code{transf} and \code{source_file}.
}
\description{
S7 generic for building the list of parameters to calibrate
from a YAML specification file.
}
\examples{
\dontrun{
mod  <- new_atlantis() |>
  atlantis_load_files(c(
    atlantis_examples("inputs", "tiny.bgm"),
    atlantis_examples("inputs", "tiny_biol.prm"),
    atlantis_examples("inputs", "tiny_groups.csv")
  ))
mod |>
  generate_calibration_table(
   atlantis_examples("calibrate", "mum.yaml")
 )
}

}
