% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlantis_methods-calibrate.R
\name{generate_calibration_table}
\alias{generate_calibration_table}
\title{Generate a calibration table}
\usage{
generate_calibration_table(x, ...)
}
\arguments{
\item{x}{An \code{Atlantis} object.}

\item{path}{Path to a YAML file describing the parameters to calibrate.}

\item{version}{Atlantis version identifier. Default is \code{"3-6722"}.}
}
\value{
A list of data frames, one per parameter entry, with columns
\code{name}, \code{min}, \code{max}, \code{position}, \code{transf} and \code{source_file}.
}
\description{
Build the list of parameters to calibrate from a YAML specification file.
}
