% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_io_prm.R
\name{read_prm}
\alias{read_prm}
\alias{write_prm}
\title{Parse .prm file contents into structured list}
\usage{
read_prm(file)

write_prm(x, file = stdout(), add_comments = TRUE, sep = " ")
}
\arguments{
\item{file}{Either a path to a file or a connection.}

\item{x}{a list of key-value pairs.}

\item{add_comments}{Should comments be added.}

\item{sep}{the separation bewteen key and value.}
}
\value{
A list made of key-value elements.
Note that comments and length of vectr

The prm file.
}
\description{
Parse .prm file contents into structured list
}
\details{
.prm files are simple key-value text files:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ #' Comments start with #
 scalar_key  1.0
 biosedprofile  gaussian
 # Arrays: key followed by count, then values on next line
 array_key 4
 1.0 1.0 1.0 1.0.
}\if{html}{\out{</div>}}

The key differences between \code{read_prm()} and the Atlantis parser are:
\itemize{
\item Comments: lines starting with # are not skipped, they are collected and stored as
an attributes to the list elements
\item Array padding: in the original parser, if fewer values than declared size,
the last value is repeated to fill. \code{read_prm()} will throw an error instead.
}

To be consistent with the Atlantis parser, lines with more than 4000
characters (a lot, even for vectors) throw an error. Also, key duplicates will
trigger warnings and the
}
\section{Functions}{
\itemize{
\item \code{write_prm()}: Write a prm file

}}
